requirements.txt
openai-whisper
google-genai
pydub
python-dotenv
Flask